Abstracted out Window into a new class
