+++ src/main/java/duke/Duke.java
@@ -1,19 +1,6 @@
 package duke;
 
 import javafx.application.Application;
-import javafx.application.Platform;
-import javafx.fxml.FXML;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.Region;
-import javafx.scene.layout.VBox;
-import javafx.scene.shape.Circle;
 import javafx.stage.Stage;
 
 /**
@@ -23,18 +10,7 @@ public class Duke extends Application {
     private Ui ui;
     private Storage storage;
     private TaskList tasks;
-    @FXML
-    private ScrollPane scrollPane;
-    @FXML
-    private VBox dialogContainer;
-    @FXML
-    private TextField userInput;
-    @FXML
-    private Button sendButton;
-    private Scene scene;
 
-    private Image user = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
-    private Image duke = new Image(this.getClass().getResourceAsStream("/images/DaDuke.png"));
     /**
      * @param filePath The file path of the txt document to read and write from.
      */
@@ -58,132 +34,23 @@ public class Duke extends Application {
     }
 
     /**
-     * Iteration 1:
-     * Creates a label with the specified text and adds it to the dialog container.
-     * @param text String containing text to add
-     * @return a label with the specified text that has word wrap enabled.
-     */
-    private Label getDialogLabel(String text) {
-        // You will need to import `javafx.scene.control.Label`.
-        Label textToAdd = new Label(text);
-        textToAdd.setWrapText(true);
-
-        return textToAdd;
-    }
-    private String getResponse(String input) {
-        return input;
-    }
-    /**
      * Iteration 2:
      * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
      * the dialog container. Clears the user input after processing.
      */
-    @FXML
-    private void handleUserInput() {
-        Label userText = new Label(userInput.getText());
-        String userInputText = userInput.getText();
-        if (userInputText.equalsIgnoreCase("Bye")) {
-            Platform.exit();
-            storage.writeToStorage();
-        }
-        System.out.println(userText.getText());
-        String responseText = this.ui.startUi(userInputText);
-        Label dukeText = new Label(getResponse(responseText));
-
-        ImageView userImageView = new ImageView(user);
-        ImageView dukeImageView = new ImageView(duke);
-
-        CircleClip clip = new CircleClip(40, 40, 40);
-        CircleClip clip2 = new CircleClip(40, 40, 40);
-        clip.clip(userImageView);
-        clip2.clip(dukeImageView);
-
-        DialogBox userDialog = DialogBox.getUserDialog(userText, userImageView, userInputText);
-        DialogBox dukeDialog = DialogBox.getDukeDialog(dukeText, dukeImageView, responseText);
-
-        dialogContainer.getChildren().addAll(
-            userDialog, dukeDialog
-        );
-        userInput.clear();
-    }
-
     @Override
     public void start(Stage stage) {
-        Label helloWorld = new Label("Hello World!"); // Creating a new Label control
-        Scene scene = new Scene(helloWorld); // Setting the scene to be our Label
 
-        stage.setScene(scene); // Setting the stage to show our screen
-        stage.show(); // Render the stage.
+        Window window = new Window(stage, storage, ui);
 
-        //The container for the content of the chat to scroll.
-        scrollPane = new ScrollPane();
-        dialogContainer = new VBox();
+        window.initializeWindow();
 
-        // Set spacing between DialogBoxes
-        dialogContainer.setSpacing(20.0);
+        window.formatWindow();
 
-        scrollPane.setContent(dialogContainer);
+        window.addUserInput();
 
-        userInput = new TextField();
-        sendButton = new Button("Send");
+        window.welcomeMessage();
 
-        AnchorPane mainLayout = new AnchorPane();
-        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
-
-        scene = new Scene(mainLayout);
-
-        stage.setScene(scene);
-        stage.show();
-
-        //Step 2. Formatting the window to look as expected
-        stage.setTitle("Duke");
-        stage.setResizable(false);
-        stage.setMinHeight(600.0);
-        stage.setMinWidth(400.0);
-
-        mainLayout.setPrefSize(400.0, 600.0);
-
-        scrollPane.setPrefSize(385, 535);
-        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
-        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
-
-        scrollPane.setVvalue(1.0);
-        scrollPane.setFitToWidth(true);
-
-        // You will need to import `javafx.scene.layout.Region` for this.
-        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
-
-        userInput.setPrefWidth(325.0);
-
-        sendButton.setPrefWidth(55.0);
-
-        AnchorPane.setTopAnchor(scrollPane, 1.0);
-
-        AnchorPane.setBottomAnchor(sendButton, 1.0);
-        AnchorPane.setRightAnchor(sendButton, 1.0);
-
-        AnchorPane.setLeftAnchor(userInput , 1.0);
-        AnchorPane.setBottomAnchor(userInput, 1.0);
-
-        //Step 3. Add functionality to handle user input.
-
-        sendButton.setOnMouseClicked((event) -> {
-            handleUserInput();
-        });
-
-        userInput.setOnAction((event) -> {
-            handleUserInput();
-        });
-
-        //Scroll down to the end every time dialogContainer's height changes.
-        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
-
-        CircleClip clip = new CircleClip(40, 40, 40);
-        ImageView dukeImageView = new ImageView(duke);
-        clip.clip(dukeImageView);
-        String welMessage = "Hello! I'm DukeBot\n" + "What can I do for you?\n";
-        Label welcomeMessage = new Label(welMessage);
-        dialogContainer.getChildren().add(DialogBox.getDukeDialog(welcomeMessage, dukeImageView, welMessage));
     }
 
 }
